+++ NEWS.md
@@ -1,3 +1,11 @@
-﻿# jab.adverse.reactions 1.0.0 (11 January 2022)
+﻿# jab.adverse.reactions 1.0.1 (13 January 2022)
+
+* Added 3 additional references
+* Added a CITATION file in the inst directory
+* Added `install.load` as a Suggests package
+* Corrected and Updated the README file & updated the DESCRIPTION file
+
+
+# jab.adverse.reactions 1.0.0 (11 January 2022)
 
 * Initial release